# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AiSOC Platform API
  description: |
    Open-source AI Security Operations Center (v5.2.0).

    Autonomous threat detection, investigation, and response — with UEBA behavioral
    analytics, honeytoken deception, purple-team adversary emulation, enterprise RBAC,
    immutable audit logging, the v5.2 Investigation Ledger (signed agent actions),
    Ambient Copilot, Responder PWA, public evaluation harness, MCP server, and
    SOC 2 / ISO 27001 / NIST CSF compliance dashboards.

    ## Authentication
    Most endpoints require a JWT Bearer token obtained from `POST /api/v1/auth/login`.
    Service-to-service calls may use an API key via `X-API-Key` header.

    ## Multi-tenancy
    All data is isolated via PostgreSQL Row-Level Security. The `tenant_id` is derived
    from the authenticated token and cannot be overridden by the caller.

    ## Rate limiting
    Default: 1000 req/min per API key; 100 req/min per user token.
    Limits are surfaced via `X-RateLimit-*` response headers.
  version: 5.2.0
  contact:
    name: AiSOC Community
    url: https://github.com/beenuar/AiSOC/security/advisories/new
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
...
paths:
...
  /api/v1/plugins:
    get:
      tags:
      - plugins
      summary: List Plugins
      description: List all currently loaded plugins, optionally filtered by type.
      operationId: list_plugins_api_v1_plugins_get
      security:
      - HTTPBearer: []
      parameters:
      - name: plugin_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Plugin Type
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PluginOut'
                title: Response List Plugins Api V1 Plugins Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/plugins/discover:
    post:
      tags:
      - plugins
      summary: Discover Plugins
      description: Re-scan the plugins directory and load any new plugins found.
      operationId: discover_plugins_api_v1_plugins_discover_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoverResponse'
      security:
      - HTTPBearer: []
  /api/v1/plugins/{plugin_id}:
    get:
      tags:
      - plugins
      summary: Get Plugin
      description: Get details for a specific plugin.
      operationId: get_plugin_api_v1_plugins__plugin_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: plugin_id
        in: path
        required: true
        schema:
          type: string
          title: Plugin Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - plugins
      summary: Unload Plugin
      description: Unload a plugin from memory (does not remove files from disk).
      operationId: unload_plugin_api_v1_plugins__plugin_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: plugin_id
        in: path
        required: true
        schema:
          type: string
          title: Plugin Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/plugins/{plugin_id}/enable:
    post:
      tags:
      - plugins
      summary: Enable Plugin
      description: Enable a loaded plugin.
      operationId: enable_plugin_api_v1_plugins__plugin_id__enable_post
      security:
      - HTTPBearer: []
      parameters:
      - name: plugin_id
        in: path
        required: true
        schema:
          type: string
          title: Plugin Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/plugins/{plugin_id}/disable:
    post:
      tags:
      - plugins
      summary: Disable Plugin
      description: Disable a loaded plugin (keeps it in memory but blocks invocations).
      operationId: disable_plugin_api_v1_plugins__plugin_id__disable_post
      security:
      - HTTPBearer: []
      parameters:
      - name: plugin_id
        in: path
        required: true
        schema:
          type: string
          title: Plugin Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/plugins/{plugin_id}/reload:
    post:
      tags:
      - plugins
      summary: Reload Plugin
      description: Unload and re-import a plugin from disk (hot-reload).
      operationId: reload_plugin_api_v1_plugins__plugin_id__reload_post
      security:
      - HTTPBearer: []
      parameters:
      - name: plugin_id
        in: path
        required: true
        schema:
          type: string
          title: Plugin Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/plugins/{plugin_id}/run:
    post:
      tags:
      - plugins
      summary: Run Plugin
      description: 'Directly invoke a plugin.

        Supports enrichers, actions, and connectors.'
      operationId: run_plugin_api_v1_plugins__plugin_id__run_post
      security:
      - HTTPBearer: []
      parameters:
      - name: plugin_id
        in: path
        required: true
        schema:
          type: string
          title: Plugin Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    DiscoverResponse:
      properties:
        discovered:
          items:
            type: string
          type: array
          title: Discovered
      type: object
      required:
      - discovered
      title: DiscoverResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PluginManifestOut:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        version:
          type: string
          title: Version
        plugin_type:
          type: string
          title: Plugin Type
        description:
          type: string
          title: Description
        author:
          type: string
          title: Author
        tags:
          items:
            type: string
          type: array
          title: Tags
        config_schema:
          additionalProperties: true
          type: object
          title: Config Schema
      type: object
      required:
      - id
      - name
      - version
      - plugin_type
      - description
      - author
      - tags
      - config_schema
      title: PluginManifestOut
    PluginOut:
      properties:
        id:
          type: string
          title: Id
        manifest:
          $ref: '#/components/schemas/PluginManifestOut'
        enabled:
          type: boolean
          title: Enabled
        loaded_at:
          type: number
          title: Loaded At
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      type: object
      required:
      - id
      - manifest
      - enabled
      - loaded_at
      - error
      title: PluginOut
...
    RunRequest:
      properties:
        payload:
          additionalProperties: true
          type: object
          title: Payload
          default: {}
        context:
          additionalProperties: true
          type: object
          title: Context
          default: {}
      type: object
      title: RunRequest
    RunResponse:
      properties:
        plugin_id:
          type: string
          title: Plugin Id
        result:
          additionalProperties: true
          type: object
          title: Result
      type: object
      required:
      - plugin_id
      - result
      title: RunResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer